Improved log customization
