+++ phase4-dbnalliance-client/src/test/java/com/helger/phase4/dbnalliance/MainPhase4DBNAllianceSenderLocalHost8080.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2015-2025 Philip Helger (www.helger.com)
+ * Copyright (C) 2024-2025 Philip Helger (www.helger.com)
  * philip[at]helger[dot]com
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
