+++ phase4-peppol-client/src/test/java/com/helger/phase4/peppol/receivers/MainPhase4PeppolSenderHelger.java
@@ -27,6 +27,7 @@ import com.helger.phase4.client.IAS4ClientBuildMessageCallback;
 import com.helger.phase4.dump.AS4DumpManager;
 import com.helger.phase4.dump.AS4IncomingDumperFileBased;
 import com.helger.phase4.dump.AS4OutgoingDumperFileBased;
+import com.helger.phase4.logging.Phase4LogCustomizer;
 import com.helger.phase4.logging.Phase4LoggerFactory;
 import com.helger.phase4.model.message.AS4UserMessage;
 import com.helger.phase4.model.message.AbstractAS4Message;
@@ -52,67 +53,66 @@ public final class MainPhase4PeppolSenderHelger
 
   public static void send ()
   {
-    Phase4LoggerFactory.setThreadLocalPrefix ("[PeppolSender] <");
-    Phase4LoggerFactory.setThreadLocalSuffix (">");
-
-    try
-    {
-      final Element aPayloadElement = DOMReader.readXMLDOM (new File ("src/test/resources/external/examples/base-example.xml"))
-                                               .getDocumentElement ();
-      if (aPayloadElement == null)
-        throw new IllegalStateException ("Failed to read XML file to be send");
-
-      // Start configuring here
-      final IParticipantIdentifier aReceiverID = Phase4PeppolSender.IF.createParticipantIdentifierWithDefaultScheme ("9915:helger");
-      final IAS4ClientBuildMessageCallback aBuildMessageCallback = new IAS4ClientBuildMessageCallback ()
+    Phase4LogCustomizer.runWithLogPrefixAndSuffix ("[phase4Sender] <", ">", () -> {
+      try
       {
-        public void onAS4Message (final AbstractAS4Message <?> aMsg)
+        final Element aPayloadElement = DOMReader.readXMLDOM (new File ("src/test/resources/external/examples/base-example.xml"))
+                                                 .getDocumentElement ();
+        if (aPayloadElement == null)
+          throw new IllegalStateException ("Failed to read XML file to be send");
+
+        // Start configuring here
+        final IParticipantIdentifier aReceiverID = Phase4PeppolSender.IF.createParticipantIdentifierWithDefaultScheme ("9915:helger");
+        final IAS4ClientBuildMessageCallback aBuildMessageCallback = new IAS4ClientBuildMessageCallback ()
         {
-          final AS4UserMessage aUserMsg = (AS4UserMessage) aMsg;
-          LOGGER.info ("Sending out AS4 message with message ID '" +
-                       aUserMsg.getEbms3UserMessage ().getMessageInfo ().getMessageId () +
-                       "'");
-          LOGGER.info ("Sending out AS4 message with conversation ID '" +
-                       aUserMsg.getEbms3UserMessage ().getCollaborationInfo ().getConversationId () +
-                       "'");
+          public void onAS4Message (final AbstractAS4Message <?> aMsg)
+          {
+            final AS4UserMessage aUserMsg = (AS4UserMessage) aMsg;
+            LOGGER.info ("Sending out AS4 message with message ID '" +
+                         aUserMsg.getEbms3UserMessage ().getMessageInfo ().getMessageId () +
+                         "'");
+            LOGGER.info ("Sending out AS4 message with conversation ID '" +
+                         aUserMsg.getEbms3UserMessage ().getCollaborationInfo ().getConversationId () +
+                         "'");
+          }
+        };
+        final IAS4RawResponseConsumer aRRC = x -> {
+          LOGGER.info ("Response status line: " + x.getResponseStatusLine ());
+          LOGGER.info ("Response headers:");
+          x.getResponseHeaders ().forEachSingleHeader ( (k, v) -> LOGGER.info ("  " + k + "=" + v), false);
+        };
+        final IAS4SignalMessageValidationResultHandler aSignalMsgValidationResultHdl = new LoggingAS4SignalMsgValidationResultHandler ();
+        final PeppolUserMessageBuilder aBuilder = Phase4PeppolSender.builder ()
+                                                                    .documentTypeID (Phase4PeppolSender.IF.createDocumentTypeIdentifierWithDefaultScheme ("urn:oasis:names:specification:ubl:schema:xsd:Invoice-2::Invoice##urn:cen.eu:en16931:2017#compliant#urn:fdc:peppol.eu:2017:poacc:billing:3.0::2.1"))
+                                                                    .processID (Phase4PeppolSender.IF.createProcessIdentifierWithDefaultScheme ("urn:fdc:peppol.eu:2017:poacc:billing:01:1.0"))
+                                                                    .senderParticipantID (Phase4PeppolSender.IF.createParticipantIdentifierWithDefaultScheme ("9915:phase4-test-sender"))
+                                                                    .receiverParticipantID (aReceiverID)
+                                                                    .senderPartyID ("POP000306")
+                                                                    .countryC1 ("AT")
+                                                                    .payload (aPayloadElement)
+                                                                    .smpClient (new SMPClientReadOnly (Phase4PeppolSender.URL_PROVIDER,
+                                                                                                       aReceiverID,
+                                                                                                       ESML.DIGIT_TEST))
+                                                                    .checkReceiverAPCertificate (true)
+                                                                    .disableValidation ()
+                                                                    .buildMessageCallback (aBuildMessageCallback)
+                                                                    .signalMsgValidationResultHdl (aSignalMsgValidationResultHdl)
+                                                                    .rawResponseConsumer (aRRC);
+        final EAS4UserMessageSendResult eResult = aBuilder.sendMessageAndCheckForReceipt ();
+        LOGGER.info ("Peppol send result: " + eResult);
+
+        if (eResult.isSuccess ())
+        {
+          // Remember item for reporting
+          aBuilder.createAndStorePeppolReportingItemAfterSending ("your-c1-end-user-id");
         }
-      };
-      final IAS4RawResponseConsumer aRRC = x -> {
-        LOGGER.info ("Response status line: " + x.getResponseStatusLine ());
-        LOGGER.info ("Response headers:");
-        x.getResponseHeaders ().forEachSingleHeader ( (k, v) -> LOGGER.info ("  " + k + "=" + v), false);
-      };
-      final IAS4SignalMessageValidationResultHandler aSignalMsgValidationResultHdl = new LoggingAS4SignalMsgValidationResultHandler ();
-      final PeppolUserMessageBuilder aBuilder = Phase4PeppolSender.builder ()
-                                                                  .documentTypeID (Phase4PeppolSender.IF.createDocumentTypeIdentifierWithDefaultScheme ("urn:oasis:names:specification:ubl:schema:xsd:Invoice-2::Invoice##urn:cen.eu:en16931:2017#compliant#urn:fdc:peppol.eu:2017:poacc:billing:3.0::2.1"))
-                                                                  .processID (Phase4PeppolSender.IF.createProcessIdentifierWithDefaultScheme ("urn:fdc:peppol.eu:2017:poacc:billing:01:1.0"))
-                                                                  .senderParticipantID (Phase4PeppolSender.IF.createParticipantIdentifierWithDefaultScheme ("9915:phase4-test-sender"))
-                                                                  .receiverParticipantID (aReceiverID)
-                                                                  .senderPartyID ("POP000306")
-                                                                  .countryC1 ("AT")
-                                                                  .payload (aPayloadElement)
-                                                                  .smpClient (new SMPClientReadOnly (Phase4PeppolSender.URL_PROVIDER,
-                                                                                                     aReceiverID,
-                                                                                                     ESML.DIGIT_TEST))
-                                                                  .checkReceiverAPCertificate (true)
-                                                                  .disableValidation ()
-                                                                  .buildMessageCallback (aBuildMessageCallback)
-                                                                  .signalMsgValidationResultHdl (aSignalMsgValidationResultHdl)
-                                                                  .rawResponseConsumer (aRRC);
-      final EAS4UserMessageSendResult eResult = aBuilder.sendMessageAndCheckForReceipt ();
-      LOGGER.info ("Peppol send result: " + eResult);
 
-      if (eResult.isSuccess ())
-      {
-        // Remember item for reporting
-        aBuilder.createAndStorePeppolReportingItemAfterSending ("your-c1-end-user-id");
       }
-
-    }
-    catch (final Exception ex)
-    {
-      LOGGER.error ("Error sending Peppol message via AS4", ex);
-    }
+      catch (final Exception ex)
+      {
+        LOGGER.error ("Error sending Peppol message via AS4", ex);
+      }
+    });
   }
 
   public static void main (final String [] args)
